Rename methods and preserve backward compatibility
